Move useInterpret to file
