+++ packages/xstate-react/src/useMachine.ts
@@ -1,7 +1,6 @@
 import { useState, useEffect, useRef } from 'react';
 import useIsomorphicLayoutEffect from 'use-isomorphic-layout-effect';
 import {
-  interpret,
   EventObject,
   StateMachine,
   State,
@@ -18,7 +17,7 @@ import {
   ReactActionObject,
   ReactEffectType
 } from './types';
-import useConstant from './useConstant';
+import { useInterpret } from './useInterpret';
 import { partition } from './utils';
 
 function createReactActionFunction<TContext, TEvent extends EventObject>(
@@ -71,7 +70,7 @@ function executeEffect<TContext, TEvent extends EventObject>(
   originalExec();
 }
 
-interface UseMachineOptions<TContext, TEvent extends EventObject> {
+export interface UseMachineOptions<TContext, TEvent extends EventObject> {
   /**
    * If provided, will be merged with machine's `context`.
    */
@@ -188,87 +187,4 @@ export function useMachine<
   }, [state]);
 
   return [state, service.send, service];
-}
-
-export function useInterpret<
-  TContext,
-  TEvent extends EventObject,
-  TTypestate extends Typestate<TContext> = { value: any; context: TContext }
->(
-  getMachine: MaybeLazy<StateMachine<TContext, any, TEvent, TTypestate>>,
-  options: Partial<InterpreterOptions> &
-    Partial<UseMachineOptions<TContext, TEvent>> &
-    Partial<MachineOptions<TContext, TEvent>> = {}
-): Interpreter<TContext, any, TEvent, TTypestate> {
-  const machine = useConstant(() => {
-    return typeof getMachine === 'function' ? getMachine() : getMachine;
-  });
-
-  if (
-    process.env.NODE_ENV !== 'production' &&
-    typeof getMachine !== 'function'
-  ) {
-    const [initialMachine] = useState(machine);
-
-    if (machine !== initialMachine) {
-      console.warn(
-        'Machine given to `useMachine` has changed between renders. This is not supported and might lead to unexpected results.\n' +
-          'Please make sure that you pass the same Machine as argument each time.'
-      );
-    }
-  }
-
-  const {
-    context,
-    guards,
-    actions,
-    activities,
-    services,
-    delays,
-    state: rehydratedState,
-    ...interpreterOptions
-  } = options;
-
-  const service = useConstant(() => {
-    const machineConfig = {
-      context,
-      guards,
-      actions,
-      activities,
-      services,
-      delays
-    };
-    const machineWithConfig = machine.withConfig(machineConfig, {
-      ...machine.context,
-      ...context
-    } as TContext);
-
-    return interpret(machineWithConfig, {
-      deferEvents: true,
-      ...interpreterOptions
-    });
-  });
-
-  useIsomorphicLayoutEffect(() => {
-    service.start(
-      rehydratedState ? (State.create(rehydratedState) as any) : undefined
-    );
-
-    return () => {
-      service.stop();
-    };
-  }, []);
-
-  // Make sure actions and services are kept updated when they change.
-  // This mutation assignment is safe because the service instance is only used
-  // in one place -- this hook's caller.
-  useEffect(() => {
-    Object.assign(service.machine.options.actions, actions);
-  }, [actions]);
-
-  useEffect(() => {
-    Object.assign(service.machine.options.services, services);
-  }, [services]);
-
-  return service;
 }
