+++ packages/xstate-react/test/useInterpret.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { createMachine } from 'xstate';
 import { render, cleanup, fireEvent } from '@testing-library/react';
-import { useInterpret } from '../src/useMachine';
+import { useInterpret } from '../src/useInterpret';
 
 afterEach(cleanup);
